refactor(ts): split command.ts - extract find_action_dispatch + fanout (mirror #146)
