+++ typescript/packages/core/src/workspace/executor/command.ts
@@ -31,6 +31,8 @@ import { ExecutionNode } from '../types.ts'
 import { asyncChain } from '../../io/stream.ts'
 import type { DispatchFn } from './cross_mount.ts'
 import { handleCrossMount, isCrossMount } from './cross_mount.ts'
+import { applyFindActions } from './find_action_dispatch.ts'
+import { fanOutTraversal, shouldFanOut } from './fanout.ts'
 import { maybeWithTimeout } from '../../commands/builtin/utils/safeguard.ts'
 import { resolveAcrossMounts, resolveSafeguard } from '../../commands/safeguard.ts'
 import type { ExecuteNodeFn } from './jobs.ts'
@@ -340,130 +342,6 @@ interface GuardResult {
   exitCode: number
 }
 
-const TRAVERSAL_CMDS: ReadonlySet<string> = new Set(['find', 'tree', 'du'])
-
-function fnmatch(name: string, pattern: string): boolean {
-  let re = ''
-  for (const ch of pattern) {
-    if (ch === '*') re += '.*'
-    else if (ch === '?') re += '.'
-    else if ('.+^$(){}|\\'.includes(ch)) re += '\\' + ch
-    else re += ch
-  }
-  return new RegExp('^' + re + '$').test(name)
-}
-
-async function applyFindActions(
-  stdout: ByteSource | null,
-  flagKwargs: Record<string, string | boolean>,
-  registry: MountRegistry,
-  cwd: string,
-): Promise<[ByteSource | null, Uint8Array]> {
-  const hasDelete = flagKwargs.delete === true
-  const hasPrint0 = flagKwargs.print0 === true
-  const hasLs = flagKwargs.ls === true
-  const hasPrint = flagKwargs.print === true
-
-  if (!hasDelete && !hasPrint0 && !hasLs) return [stdout, new Uint8Array()]
-  if (stdout === null) return [stdout, new Uint8Array()]
-
-  const data = await materialize(stdout)
-  const text = new TextDecoder().decode(data)
-  const matches = text.split('\n').filter((p) => p !== '')
-  const errors: Uint8Array[] = []
-  const enc = new TextEncoder()
-
-  let outputMatches: string[]
-
-  if (hasDelete) {
-    const deletable = matches.filter((p) => !registry.isMountRoot(p))
-    const ordered = [...deletable].sort(
-      (a, b) => (b.match(/\//g) ?? []).length - (a.match(/\//g) ?? []).length,
-    )
-    for (const path of ordered) {
-      const mount = registry.mountFor(path)
-      if (mount === null) {
-        errors.push(enc.encode(`find: cannot delete '${path}': no mount\n`))
-        continue
-      }
-      const slash = path.lastIndexOf('/')
-      const ps = new PathSpec({
-        original: path,
-        directory: slash >= 0 ? path.slice(0, slash + 1) : '/',
-        resolved: true,
-      })
-      try {
-        const [, rmIo] = await mount.executeCmd('rm', [ps], [], {}, { stdin: null, cwd })
-        if (rmIo.exitCode !== 0) {
-          const errBytes = await materialize(rmIo.stderr)
-          if (errBytes.length > 0) errors.push(errBytes)
-          else errors.push(enc.encode(`find: cannot delete '${path}'\n`))
-        }
-      } catch (err) {
-        const msg = err instanceof Error ? err.message : String(err)
-        errors.push(enc.encode(`find: cannot delete '${path}': ${msg}\n`))
-      }
-    }
-    outputMatches = hasPrint ? matches : []
-  } else if (hasLs) {
-    outputMatches = []
-    for (const path of matches) {
-      const mount = registry.mountFor(path)
-      if (mount === null) {
-        errors.push(enc.encode(`find: cannot ls '${path}': no mount\n`))
-        continue
-      }
-      const slash = path.lastIndexOf('/')
-      const ps = new PathSpec({
-        original: path,
-        directory: slash >= 0 ? path.slice(0, slash + 1) : '/',
-        resolved: true,
-      })
-      try {
-        const [lsOut] = await mount.executeCmd(
-          'ls',
-          [ps],
-          [],
-          { args_l: true, d: true },
-          { stdin: null, cwd },
-        )
-        if (lsOut !== null) {
-          const line = new TextDecoder().decode(await materialize(lsOut)).replace(/\n+$/, '')
-          if (line !== '') outputMatches.push(line)
-        }
-      } catch (err) {
-        const msg = err instanceof Error ? err.message : String(err)
-        errors.push(enc.encode(`find: cannot ls '${path}': ${msg}\n`))
-      }
-    }
-  } else {
-    outputMatches = matches
-  }
-
-  const errBlob = errors.length > 0 ? concat(errors) : new Uint8Array()
-  if (outputMatches.length === 0) return [null, errBlob]
-
-  let body: Uint8Array
-  if (hasPrint0) {
-    body = enc.encode(outputMatches.join('\x00') + '\x00')
-  } else {
-    body = enc.encode(outputMatches.join('\n') + '\n')
-  }
-  return [body, errBlob]
-}
-
-function concat(parts: readonly Uint8Array[]): Uint8Array {
-  let total = 0
-  for (const p of parts) total += p.length
-  const out = new Uint8Array(total)
-  let off = 0
-  for (const p of parts) {
-    out.set(p, off)
-    off += p.length
-  }
-  return out
-}
-
 function checkMountRootGuard(
   cmdName: string,
   paths: readonly PathSpec[],
@@ -539,230 +417,6 @@ function checkMountRootGuard(
   }
 
   return null
-}
-
-function pathSegments(path: string): string[] {
-  return path.split('/').filter((s) => s !== '')
-}
-
-function shouldFanOut(
-  cmdName: string,
-  paths: readonly PathSpec[],
-  flagKwargs: Record<string, string | boolean>,
-  registry: MountRegistry,
-): boolean {
-  if (paths.length === 0 || paths[0] === undefined) return false
-  if (registry.descendantMounts(paths[0].original).length === 0) return false
-  if (TRAVERSAL_CMDS.has(cmdName)) return true
-  if (cmdName === 'grep') {
-    return flagKwargs.r === true || flagKwargs.R === true || flagKwargs.recursive === true
-  }
-  if (cmdName === 'ls') {
-    return flagKwargs.R === true
-  }
-  return false
-}
-
-function adjustDepthFlags(
-  flagKwargs: Record<string, string | boolean>,
-  parentPath: string,
-  mountPrefix: string,
-): Record<string, string | boolean> | null {
-  const parentDepth = pathSegments(parentPath).length
-  const mountDepth = pathSegments(mountPrefix).length
-  const delta = mountDepth - parentDepth
-  const out: Record<string, string | boolean> = { ...flagKwargs }
-  if ('maxdepth' in out) {
-    const orig = Number(out.maxdepth)
-    if (!Number.isNaN(orig)) {
-      const md = orig - delta
-      if (md < 0) return null
-      out.maxdepth = String(md)
-    }
-  }
-  if ('mindepth' in out) {
-    const orig = Number(out.mindepth)
-    if (!Number.isNaN(orig)) {
-      out.mindepth = String(Math.max(0, orig - delta))
-    }
-  }
-  return out
-}
-
-function synthesizeFindMountEntries(
-  targetPath: string,
-  descendants: readonly Mount[],
-  flagKwargs: Record<string, string | boolean>,
-): string {
-  const typeFilter = flagKwargs.type
-  if (typeFilter !== undefined && typeFilter !== 'd') return ''
-  const parentDepth = pathSegments(targetPath).length
-  const maxRaw = flagKwargs.maxdepth
-  const minRaw = flagKwargs.mindepth
-  const maxDepth = maxRaw !== undefined ? Number(maxRaw) : null
-  const minDepth = minRaw !== undefined ? Number(minRaw) : 0
-  const namePat = typeof flagKwargs.name === 'string' ? flagKwargs.name : null
-  const inamePat = typeof flagKwargs.iname === 'string' ? flagKwargs.iname : null
-  const out: string[] = []
-  for (const m of descendants) {
-    const prefixNoSlash = m.prefix.replace(/\/+$/, '')
-    const depth = pathSegments(prefixNoSlash).length - parentDepth
-    if (depth < minDepth) continue
-    if (maxDepth !== null && !Number.isNaN(maxDepth) && depth > maxDepth) continue
-    const segs = prefixNoSlash.split('/').filter((s) => s !== '')
-    const base = segs[segs.length - 1] ?? prefixNoSlash
-    if (namePat !== null && !fnmatch(base, namePat)) continue
-    if (inamePat !== null && !fnmatch(base.toLowerCase(), inamePat.toLowerCase())) continue
-    out.push(prefixNoSlash)
-  }
-  return out.join('\n')
-}
-
-async function filterUnderPrefixes(
-  stdout: ByteSource,
-  descendantPrefixes: readonly string[],
-): Promise<Uint8Array> {
-  const data = await materialize(stdout)
-  const text = new TextDecoder().decode(data)
-  const outLines: string[] = []
-  for (const line of text.split('\n')) {
-    if (line === '') continue
-    let path = line
-    for (const sep of ['\t', ':']) {
-      const idx = path.indexOf(sep)
-      if (idx >= 0) {
-        path = path.slice(0, idx)
-        break
-      }
-    }
-    if (path.startsWith('/')) {
-      let shadowed = false
-      for (const pre of descendantPrefixes) {
-        if (path === pre || path.startsWith(pre + '/')) {
-          shadowed = true
-          break
-        }
-      }
-      if (shadowed) continue
-    }
-    outLines.push(line)
-  }
-  if (outLines.length === 0) return new Uint8Array()
-  return new TextEncoder().encode(outLines.join('\n') + '\n')
-}
-
-async function fanOutTraversal(
-  cmdName: string,
-  paths: readonly PathSpec[],
-  texts: readonly string[],
-  flagKwargs: Record<string, string | boolean>,
-  registry: MountRegistry,
-  primaryMount: Mount,
-  cwd: string,
-  cmdStr: string,
-  stdin: ByteSource | null,
-  ensureOpen: ((resource: Resource) => Promise<void>) | undefined,
-): Promise<Result> {
-  const targetPath = paths[0]?.original ?? cwd
-  const descendants = registry.descendantMounts(targetPath)
-  const descendantPrefixes = descendants.map((m) => m.prefix.replace(/\/+$/, ''))
-
-  const allStdout: Uint8Array[] = []
-  let mergedIo = new IOResult()
-  let finalExit = 0
-  let successSeen = false
-
-  const mountsToRun: Mount[] = [primaryMount, ...descendants]
-  for (const mount of mountsToRun) {
-    let subPaths: PathSpec[]
-    let subFlags: Record<string, string | boolean>
-    if (mount === primaryMount) {
-      subPaths = [...paths]
-      subFlags = { ...flagKwargs }
-    } else {
-      const adjusted = adjustDepthFlags(flagKwargs, targetPath, mount.prefix)
-      if (adjusted === null) continue
-      subFlags = adjusted
-      const mountRoot = mount.prefix.replace(/\/+$/, '') || '/'
-      subPaths = [
-        new PathSpec({
-          original: mountRoot,
-          directory: mountRoot,
-          prefix: mount.prefix.replace(/\/+$/, ''),
-        }),
-      ]
-    }
-    if (ensureOpen !== undefined) {
-      try {
-        await ensureOpen(mount.resource)
-      } catch {
-        continue
-      }
-    }
-    let stdout: ByteSource | null
-    let io: IOResult
-    try {
-      const result = await mount.executeCmd(cmdName, subPaths, [...texts], subFlags, {
-        stdin,
-        cwd,
-      })
-      stdout = result[0]
-      io = result[1]
-    } catch {
-      continue
-    }
-    if (mount === primaryMount && descendantPrefixes.length > 0 && stdout !== null) {
-      stdout = await filterUnderPrefixes(stdout, descendantPrefixes)
-    }
-    if (stdout !== null) {
-      const data = await materialize(stdout)
-      if (data.length > 0) allStdout.push(data)
-    }
-    if (io.exitCode === 0) {
-      successSeen = true
-    } else if (finalExit === 0) {
-      finalExit = io.exitCode
-    }
-    mergedIo = await mergedIo.merge(io)
-  }
-
-  if (cmdName === 'find') {
-    const synthetic = synthesizeFindMountEntries(targetPath, descendants, flagKwargs)
-    if (synthetic !== '') allStdout.push(new TextEncoder().encode(synthetic))
-  }
-
-  let finalIoExit = successSeen ? 0 : finalExit
-  let combined: ByteSource | null = null
-  if (allStdout.length > 0) {
-    const parts = allStdout.map((d) => {
-      const s = new TextDecoder().decode(d).replace(/\n+$/, '')
-      return s
-    })
-    combined = new TextEncoder().encode(parts.filter((s) => s !== '').join('\n') + '\n')
-  }
-
-  if (cmdName === 'find') {
-    const [newCombined, actionErr] = await applyFindActions(combined, flagKwargs, registry, cwd)
-    combined = newCombined
-    if (actionErr.length > 0) {
-      const existing = await materialize(mergedIo.stderr)
-      const merged = new Uint8Array(existing.length + actionErr.length)
-      merged.set(existing, 0)
-      merged.set(actionErr, existing.length)
-      mergedIo.stderr = merged
-      if (finalIoExit === 0) finalIoExit = 1
-    }
-  }
-
-  mergedIo.exitCode = finalIoExit
-  mergedIo.safeguard = resolveAcrossMounts(cmdName, mountsToRun)
-  const stderrBytes = await materialize(mergedIo.stderr)
-  const exec = new ExecutionNode({
-    command: cmdStr,
-    stderr: stderrBytes,
-    exitCode: finalIoExit,
-  })
-  return [combined, mergedIo, exec]
 }
 
 async function injectChildMounts(
